Remove old and modified libs, switch to ESPAsyncWebserver, add support for ESP32-H2 and multiple Ethernet modules (#455)
* Asyncwebserver

* Squashed commit of the following:

commit 575ef02f593918ec6654c87407a4d11fc17071b8
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 17:56:11 2024 +0200

    merge master

commit 35e5adf4ecd80f9829e8801181f35dd2c1d94759
Merge: a2cc7be2 21adca01
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 17:41:04 2024 +0200

    Merge branch 'master' of github.com:technyon/nuki_hub into DM9051

commit a2cc7be2954cbd8767ab8186296c0b14134d1d0b
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 10:51:50 2024 +0200

    update nuki ble

commit 20c809f3dca28b29b219d1ff3a183f1981316de5
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 10:44:46 2024 +0200

    backup

commit dd41c218efb5270f5efeb734e64dff695920db16
Merge: 153000b5 e84b944a
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 10:40:03 2024 +0200

    Merge branch 'master' of github.com:technyon/nuki_hub into DM9051

commit 153000b5b1af7df1fbeb5263df94eb26f689cc0a
Author: technyon <j.o.schuemann@gmx.de>
Date:   Mon Aug 12 10:23:07 2024 +0200

    fix linker error

commit a93bbfbfc4301e46ff3696a763dd13c6c89efefb
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sun Aug 11 11:27:07 2024 +0200

    backup

commit f611c75ce8c35f829bcad6cf7e86188f4b3ec331
Merge: f1964917 063fbab6
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sun Aug 11 11:24:47 2024 +0200

    merge master

commit f1964917b4dade3920f1ecdb699c58630199e6da
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sat Aug 10 15:17:45 2024 +0200

    update platformio.ini

commit f448e5e8a7e93be38e09e2ab0b622199a3721af6
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sat Aug 10 11:28:09 2024 +0200

    add SPIClass instance for DM9051

commit 1f190e9aa08033535a2eb442a92e6e20409bbda1
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sat Aug 10 11:22:26 2024 +0200

    add definitions and constructor for DM9051

commit 726b3602ae91594ee1210ad5b6714f75cc5e42a7
Merge: 50a2eb13 4af90cbc
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sat Aug 10 10:19:34 2024 +0200

    merge master

commit 50a2eb136d75d90921f1c6974f18bc107bddc123
Author: technyon <j.o.schuemann@gmx.de>
Date:   Fri Aug 9 11:52:09 2024 +0200

    add comment

commit 9437e485cae169efdf8e5a7bf188a1c7e792d1e5
Author: technyon <j.o.schuemann@gmx.de>
Date:   Sun Aug 4 08:29:21 2024 +0200

    move LAN8720 definitions to seperate file

* Remove Core 2 Ethernet library

* Custom Ethernet

* GPIO and Preferences

* H2

+++ pio_package_pre.py
@@ -12,23 +12,52 @@ def recursive_purge(dir, pattern):
 
 regex = r"\#define NUKI_HUB_DATE \"(.*)\""
 content_new = ""
+file_content = ""
 
 with open ('src/Config.h', 'r' ) as readfile:
     file_content = readfile.read()
     content_new = re.sub(regex, "#define NUKI_HUB_DATE \"" + datetime.now(timezone.utc).strftime("%Y-%m-%d") + "\"", file_content, flags = re.M)
 
-with open('src/Config.h', 'w') as writefile:
-    writefile.write(content_new)
+if content_new != file_content:
+    with open('src/Config.h', 'w') as writefile:
+        writefile.write(content_new)
 
 recursive_purge("managed_components", ".component_hash")
 
-if env.get('BOARD_MCU') == "esp32":
-  board = "esp32dev"
-else:
-  board = env.get('BOARD_MCU')
+board = env.get('BOARD_MCU')
 
 if os.path.exists("sdkconfig." + board):
-  os.remove("sdkconfig." + board)
+  f1 = 0;
+  f2 = 0;
+  f3 = 0;
+  f4 = os.path.getmtime("sdkconfig." + board)
+  
+  if os.path.exists("sdkconfig.defaults." + board):
+    f1 = os.path.getmtime("sdkconfig.defaults." + board)
+    
+  if os.path.exists("sdkconfig.release.defaults"):
+    f2 = os.path.getmtime("sdkconfig.release.defaults")
+  
+  if os.path.exists("sdkconfig.defaults"):
+    f3 = os.path.getmtime("sdkconfig.defaults")
+
+  if(f1 > f4 or f2 > f4 or f3 > f4):
+    os.remove("sdkconfig." + board)
   
 if os.path.exists("sdkconfig." + board + "_dbg"):
-  os.remove("sdkconfig." + board + "_dbg")
+  f1 = 0;
+  f2 = 0;
+  f3 = 0;
+  f4 = os.path.getmtime("sdkconfig." + board + "_dbg")
+  
+  if os.path.exists("sdkconfig.defaults." + board):
+    f1 = os.path.getmtime("sdkconfig.defaults." + board)
+    
+  if os.path.exists("sdkconfig.debug.defaults"):
+    f2 = os.path.getmtime("sdkconfig.debug.defaults")
+  
+  if os.path.exists("sdkconfig.defaults"):
+    f3 = os.path.getmtime("sdkconfig.defaults")
+
+  if(f1 > f4 or f2 > f4 or f3 > f4):
+    os.remove("sdkconfig." + board + "_dbg")
